Create nonlabor_budget.txt
The template for the NONLABOR_BUDGET.tex file that gets generated by WorPT.  A lot of placeholders are in this template that get filled out by WorPT when it is processing the actual file that generates this budget table.  One can download this file, alter as desired, and move to one's own cloud account and point the WorPT script to read that file instead of this one, if fundamental format changes are desired.